# asset_generation/web/frontend/scripts/fixtures/openapi.cached.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blobert Asset Editor API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/registry/model/load_existing/open": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Open Load Existing",
        "operationId": "open_load_existing_api_registry_model_load_existing_open_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadExistingOpenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/registry/model/enemies/{family}/versions/{version_id}": {
      "patch": {
        "tags": [
          "registry"
        ],
        "summary": "Patch Enemy Version Endpoint",
        "operationId": "patch_enemy_version_endpoint_api_registry_model_enemies__family__versions__version_id__patch",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Family"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnemyVersionPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "registry"
        ],
        "summary": "Delete Enemy Version Endpoint",
        "operationId": "delete_enemy_version_endpoint_api_registry_model_enemies__family__versions__version_id__delete",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Family"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnemyVersionDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/registry/model/player_active_visual": {
      "delete": {
        "tags": [
          "registry"
        ],
        "summary": "Delete Player Active Visual Endpoint",
        "operationId": "delete_player_active_visual_endpoint_api_registry_model_player_active_visual_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerActiveVisualDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "registry"
        ],
        "summary": "Patch Player Active Visual Endpoint",
        "operationId": "patch_player_active_visual_endpoint_api_registry_model_player_active_visual_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerVisualPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/registry/model/enemies/{family}/slots": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Get Enemy Slots Endpoint",
        "operationId": "get_enemy_slots_endpoint_api_registry_model_enemies__family__slots_get",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Family"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "registry"
        ],
        "summary": "Put Enemy Slots Endpoint",
        "operationId": "put_enemy_slots_endpoint_api_registry_model_enemies__family__slots_put",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Family"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnemySlotsPut"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/registry/model/enemies/{family}/sync_animated_exports": {
      "post": {
        "tags": [
          "registry"
        ],
        "summary": "Post Sync Animated Exports Endpoint",
        "description": "Register on-disk ``animated_exports/{family}_animated_*.glb`` files missing from the manifest.",
        "operationId": "post_sync_animated_exports_endpoint_api_registry_model_enemies__family__sync_animated_exports_post",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Family"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/registry/model/player/slots": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Get Player Slots Endpoint",
        "operationId": "get_player_slots_endpoint_api_registry_model_player_slots_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "registry"
        ],
        "summary": "Put Player Slots Endpoint",
        "operationId": "put_player_slots_endpoint_api_registry_model_player_slots_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnemySlotsPut"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/registry/model/spawn_eligible/{family}": {
      "get": {
        "tags": [
          "registry"
        ],
        "summary": "Get Spawn Eligible",
        "description": "Consumer-facing view: paths eligible for default spawn pool (MRVC-4).",
        "operationId": "get_spawn_eligible_api_registry_model_spawn_eligible__family__get",
        "parameters": [
          {
            "name": "family",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Family"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "EnemySlotsPut": {
        "properties": {
          "version_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Version Ids",
            "description": "Ordered slot IDs for a family."
          }
        },
        "type": "object",
        "title": "EnemySlotsPut"
      },
      "EnemyVersionDeleteRequest": {
        "properties": {
          "delete_files": {
            "type": "boolean",
            "title": "Delete Files",
            "default": false
          },
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "default": false
          },
          "confirm_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm Text"
          },
          "target_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Path"
          }
        },
        "type": "object",
        "title": "EnemyVersionDeleteRequest"
      },
      "EnemyVersionPatch": {
        "properties": {
          "draft": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft",
            "description": "Draft flag; draft entries are not spawn-eligible."
          },
          "in_use": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Use",
            "description": "In-use (spawn pool) flag; ignored if draft."
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Optional display name; omit for no change, null or empty string clears."
          }
        },
        "type": "object",
        "title": "EnemyVersionPatch"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoadExistingOpenRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "One of enemy|player|path for constrained load-open."
          },
          "family": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family"
          },
          "version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Id"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "LoadExistingOpenRequest"
      },
      "PlayerActiveVisualDeleteRequest": {
        "properties": {
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "default": false
          }
        },
        "type": "object",
        "title": "PlayerActiveVisualDeleteRequest"
      },
      "PlayerVisualPatch": {
        "properties": {
          "draft": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path",
            "description": "Allowlisted path under python_root."
          }
        },
        "type": "object",
        "title": "PlayerVisualPatch"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}